+++ samples/lib/SignalingMsgHandler.c
@@ -1,0 +1,382 @@
+#include "../Samples.h"
+
+STATUS handleAnswer(PSampleConfiguration pSampleConfiguration, PSampleStreamingSession pSampleStreamingSession, PSignalingMessage pSignalingMessage)
+{
+    UNUSED_PARAM(pSampleConfiguration);
+    STATUS retStatus = STATUS_SUCCESS;
+    PRtcSessionDescriptionInit pAnswerSessionDescriptionInit = NULL;
+
+    pAnswerSessionDescriptionInit = (PRtcSessionDescriptionInit) MEMCALLOC(1, SIZEOF(RtcSessionDescriptionInit));
+
+    CHK_STATUS(deserializeSessionDescriptionInit(pSignalingMessage->payload, pSignalingMessage->payloadLen, pAnswerSessionDescriptionInit));
+    CHK_STATUS(setRemoteDescription(pSampleStreamingSession->pPeerConnection, pAnswerSessionDescriptionInit));
+
+    // The audio video receive routine should be per streaming session
+    if (pSampleConfiguration->receiveAudioVideoSource != NULL) {
+        THREAD_CREATE(&pSampleStreamingSession->receiveAudioVideoSenderTid, pSampleConfiguration->receiveAudioVideoSource,
+                      (PVOID) pSampleStreamingSession);
+    }
+CleanUp:
+
+    if (pAnswerSessionDescriptionInit != NULL) {
+        SAFE_MEMFREE(pAnswerSessionDescriptionInit);
+    }
+
+    CHK_LOG_ERR(retStatus);
+
+    return retStatus;
+}
+
+STATUS handleOffer(PSampleConfiguration pSampleConfiguration, PSampleStreamingSession pSampleStreamingSession, PSignalingMessage pSignalingMessage)
+{
+    STATUS retStatus = STATUS_SUCCESS;
+    PRtcSessionDescriptionInit pOfferSessionDescriptionInit = NULL;
+    NullableBool canTrickle;
+    BOOL mediaThreadStarted;
+
+    CHK(pSampleConfiguration != NULL && pSignalingMessage != NULL, STATUS_NULL_ARG);
+
+    pOfferSessionDescriptionInit = (PRtcSessionDescriptionInit) MEMCALLOC(1, SIZEOF(RtcSessionDescriptionInit));
+    MEMSET(&pSampleStreamingSession->answerSessionDescriptionInit, 0x00, SIZEOF(RtcSessionDescriptionInit));
+    DLOGD("**offer:%s", pSignalingMessage->payload);
+    CHK_STATUS(deserializeSessionDescriptionInit(pSignalingMessage->payload, pSignalingMessage->payloadLen, pOfferSessionDescriptionInit));
+    CHK_STATUS(setRemoteDescription(pSampleStreamingSession->pPeerConnection, pOfferSessionDescriptionInit));
+    canTrickle = canTrickleIceCandidates(pSampleStreamingSession->pPeerConnection);
+    /* cannot be null after setRemoteDescription */
+    CHECK(!NULLABLE_CHECK_EMPTY(canTrickle));
+    pSampleStreamingSession->remoteCanTrickleIce = canTrickle.value;
+    CHK_STATUS(setLocalDescription(pSampleStreamingSession->pPeerConnection, &pSampleStreamingSession->answerSessionDescriptionInit));
+
+    /*
+     * If remote support trickle ice, send answer now. Otherwise answer will be sent once ice candidate gathering is complete.
+     */
+    if (pSampleStreamingSession->remoteCanTrickleIce) {
+        CHK_STATUS(createAnswer(pSampleStreamingSession->pPeerConnection, &pSampleStreamingSession->answerSessionDescriptionInit));
+        CHK_STATUS(respondWithAnswer(pSampleStreamingSession));
+    }
+
+    mediaThreadStarted = ATOMIC_EXCHANGE_BOOL(&pSampleConfiguration->mediaThreadStarted, TRUE);
+    if (!mediaThreadStarted) {
+        THREAD_CREATE_WITH_PARAMS(&pSampleConfiguration->mediaSenderTid, mediaSenderRoutine, KVS_MINIMUM_THREAD_STACK_SIZE,
+                                  (PVOID) pSampleConfiguration);
+    }
+
+    // The audio video receive routine should be per streaming session
+    if (pSampleConfiguration->receiveAudioVideoSource != NULL) {
+        THREAD_CREATE(&pSampleStreamingSession->receiveAudioVideoSenderTid, pSampleConfiguration->receiveAudioVideoSource,
+                      (PVOID) pSampleStreamingSession);
+    }
+CleanUp:
+    if (pOfferSessionDescriptionInit != NULL) {
+        SAFE_MEMFREE(pOfferSessionDescriptionInit);
+    }
+
+    CHK_LOG_ERR(retStatus);
+
+    return retStatus;
+}
+
+STATUS sendSignalingMessage(PSampleStreamingSession pSampleStreamingSession, PSignalingMessage pMessage)
+{
+    STATUS retStatus = STATUS_SUCCESS;
+    BOOL locked = FALSE;
+    PSampleConfiguration pSampleConfiguration;
+    // Validate the input params
+    CHK(pSampleStreamingSession != NULL && pSampleStreamingSession->pSampleConfiguration != NULL && pMessage != NULL, STATUS_NULL_ARG);
+
+    pSampleConfiguration = pSampleStreamingSession->pSampleConfiguration;
+
+    CHK(IS_VALID_MUTEX_VALUE(pSampleConfiguration->signalingSendMessageLock) &&
+            IS_VALID_SIGNALING_CLIENT_HANDLE(pSampleConfiguration->signalingClientHandle),
+        STATUS_INVALID_OPERATION);
+
+    MUTEX_LOCK(pSampleConfiguration->signalingSendMessageLock);
+    locked = TRUE;
+    CHK_STATUS(signalingClientSendMessageSync(pSampleConfiguration->signalingClientHandle, pMessage));
+    if (pMessage->messageType == SIGNALING_MESSAGE_TYPE_ANSWER) {
+        CHK_STATUS(signalingClientGetMetrics(pSampleConfiguration->signalingClientHandle, &pSampleConfiguration->signalingClientMetrics));
+        DLOGP("[Signaling offer received to answer sent time] %" PRIu64 " ms",
+              pSampleConfiguration->signalingClientMetrics.signalingClientStats.offerToAnswerTime);
+    }
+
+CleanUp:
+
+    if (locked) {
+        MUTEX_UNLOCK(pSampleStreamingSession->pSampleConfiguration->signalingSendMessageLock);
+    }
+
+    CHK_LOG_ERR(retStatus);
+    return retStatus;
+}
+
+STATUS respondWithAnswer(PSampleStreamingSession pSampleStreamingSession)
+{
+    STATUS retStatus = STATUS_SUCCESS;
+    SignalingMessage message;
+    UINT32 buffLen = MAX_SIGNALING_MESSAGE_LEN;
+
+    CHK_STATUS(serializeSessionDescriptionInit(&pSampleStreamingSession->answerSessionDescriptionInit, message.payload, &buffLen));
+
+    message.version = SIGNALING_MESSAGE_CURRENT_VERSION;
+    message.messageType = SIGNALING_MESSAGE_TYPE_ANSWER;
+    STRNCPY(message.peerClientId, pSampleStreamingSession->peerId, MAX_SIGNALING_CLIENT_ID_LEN);
+    message.payloadLen = (UINT32) STRLEN(message.payload);
+    // SNPRINTF appends null terminator, so we do not manually add it
+    SNPRINTF(message.correlationId, MAX_CORRELATION_ID_LEN, "%llu_%llu", GETTIME(), ATOMIC_INCREMENT(&pSampleStreamingSession->correlationIdPostFix));
+    DLOGD("Responding With Answer With correlationId: %s", message.correlationId);
+    CHK_STATUS(sendSignalingMessage(pSampleStreamingSession, &message));
+
+CleanUp:
+
+    CHK_LOG_ERR(retStatus);
+    return retStatus;
+}
+
+VOID onIceCandidateHandler(UINT64 customData, PCHAR candidateJson)
+{
+    STATUS retStatus = STATUS_SUCCESS;
+    PSampleStreamingSession pSampleStreamingSession = (PSampleStreamingSession) customData;
+    SignalingMessage message;
+
+    CHK(pSampleStreamingSession != NULL, STATUS_NULL_ARG);
+
+    if (candidateJson == NULL) {
+        DLOGD("ice candidate gathering finished");
+        ATOMIC_STORE_BOOL(&pSampleStreamingSession->candidateGatheringDone, TRUE);
+
+        // if application is master and non-trickle ice, send answer now.
+        if (pSampleStreamingSession->pSampleConfiguration->channelInfo.channelRoleType == SIGNALING_CHANNEL_ROLE_TYPE_MASTER &&
+            !pSampleStreamingSession->remoteCanTrickleIce) {
+            CHK_STATUS(createAnswer(pSampleStreamingSession->pPeerConnection, &pSampleStreamingSession->answerSessionDescriptionInit));
+            CHK_STATUS(respondWithAnswer(pSampleStreamingSession));
+        } else if (pSampleStreamingSession->pSampleConfiguration->channelInfo.channelRoleType == SIGNALING_CHANNEL_ROLE_TYPE_VIEWER &&
+                   !pSampleStreamingSession->pSampleConfiguration->trickleIce) {
+            CVAR_BROADCAST(pSampleStreamingSession->pSampleConfiguration->cvar);
+        }
+
+    } else if (pSampleStreamingSession->remoteCanTrickleIce && ATOMIC_LOAD_BOOL(&pSampleStreamingSession->peerIdReceived)) {
+        message.version = SIGNALING_MESSAGE_CURRENT_VERSION;
+        message.messageType = SIGNALING_MESSAGE_TYPE_ICE_CANDIDATE;
+        STRNCPY(message.peerClientId, pSampleStreamingSession->peerId, MAX_SIGNALING_CLIENT_ID_LEN);
+        message.payloadLen = (UINT32) STRNLEN(candidateJson, MAX_SIGNALING_MESSAGE_LEN);
+        STRNCPY(message.payload, candidateJson, message.payloadLen);
+        message.correlationId[0] = '\0';
+        CHK_STATUS(sendSignalingMessage(pSampleStreamingSession, &message));
+    }
+
+CleanUp:
+
+    CHK_LOG_ERR(retStatus);
+}
+
+STATUS submitPendingIceCandidate(PPendingMessageQueue pPendingMessageQueue, PSampleStreamingSession pSampleStreamingSession)
+{
+    STATUS retStatus = STATUS_SUCCESS;
+    BOOL noPendingSignalingMessageForClient = FALSE;
+    PReceivedSignalingMessage pReceivedSignalingMessage = NULL;
+    UINT64 hashValue;
+
+    CHK(pPendingMessageQueue != NULL && pPendingMessageQueue->messageQueue != NULL && pSampleStreamingSession != NULL, STATUS_NULL_ARG);
+
+    do {
+        CHK_STATUS(stackQueueIsEmpty(pPendingMessageQueue->messageQueue, &noPendingSignalingMessageForClient));
+        if (!noPendingSignalingMessageForClient) {
+            hashValue = 0;
+            CHK_STATUS(stackQueueDequeue(pPendingMessageQueue->messageQueue, &hashValue));
+            pReceivedSignalingMessage = (PReceivedSignalingMessage) hashValue;
+            CHK(pReceivedSignalingMessage != NULL, STATUS_INTERNAL_ERROR);
+            if (pReceivedSignalingMessage->signalingMessage.messageType == SIGNALING_MESSAGE_TYPE_ICE_CANDIDATE) {
+                CHK_STATUS(handleRemoteCandidate(pSampleStreamingSession, &pReceivedSignalingMessage->signalingMessage));
+            }
+            SAFE_MEMFREE(pReceivedSignalingMessage);
+        }
+    } while (!noPendingSignalingMessageForClient);
+
+    CHK_STATUS(freeMessageQueue(pPendingMessageQueue));
+
+CleanUp:
+
+    SAFE_MEMFREE(pReceivedSignalingMessage);
+    CHK_LOG_ERR(retStatus);
+    return retStatus;
+}
+
+STATUS signalingMessageReceived(UINT64 customData, PReceivedSignalingMessage pReceivedSignalingMessage)
+{
+    STATUS retStatus = STATUS_SUCCESS;
+    PSampleConfiguration pSampleConfiguration = (PSampleConfiguration) customData;
+    BOOL peerConnectionFound = FALSE, locked = FALSE, startStats = FALSE, freeStreamingSession = FALSE;
+    UINT32 clientIdHash;
+    UINT64 hashValue = 0;
+    PPendingMessageQueue pPendingMessageQueue = NULL;
+    PSampleStreamingSession pSampleStreamingSession = NULL;
+    PReceivedSignalingMessage pReceivedSignalingMessageCopy = NULL;
+
+    CHK(pSampleConfiguration != NULL, STATUS_NULL_ARG);
+
+    MUTEX_LOCK(pSampleConfiguration->sampleConfigurationObjLock);
+    locked = TRUE;
+
+    clientIdHash = COMPUTE_CRC32((PBYTE) pReceivedSignalingMessage->signalingMessage.peerClientId,
+                                 (UINT32) STRLEN(pReceivedSignalingMessage->signalingMessage.peerClientId));
+    CHK_STATUS(hashTableContains(pSampleConfiguration->pRtcPeerConnectionForRemoteClient, clientIdHash, &peerConnectionFound));
+    if (peerConnectionFound) {
+        CHK_STATUS(hashTableGet(pSampleConfiguration->pRtcPeerConnectionForRemoteClient, clientIdHash, &hashValue));
+        pSampleStreamingSession = (PSampleStreamingSession) hashValue;
+    }
+
+    switch (pReceivedSignalingMessage->signalingMessage.messageType) {
+        case SIGNALING_MESSAGE_TYPE_OFFER:
+            // Check if we already have an ongoing master session with the same peer
+            CHK_ERR(!peerConnectionFound, STATUS_INVALID_OPERATION, "Peer connection %s is in progress",
+                    pReceivedSignalingMessage->signalingMessage.peerClientId);
+
+            /*
+             * Create new streaming session for each offer, then insert the client id and streaming session into
+             * pRtcPeerConnectionForRemoteClient for subsequent ice candidate messages. Lastly check if there is
+             * any ice candidate messages queued in pPendingSignalingMessageForRemoteClient. If so then submit
+             * all of them.
+             */
+
+            if (pSampleConfiguration->streamingSessionCount == ARRAY_SIZE(pSampleConfiguration->sampleStreamingSessionList)) {
+                DLOGW("Max simultaneous streaming session count reached.");
+
+                // Need to remove the pending queue if any.
+                // This is a simple optimization as the session cleanup will
+                // handle the cleanup of pending message queue after a while
+                CHK_STATUS(getPendingMessageQueueForHash(pSampleConfiguration->pPendingSignalingMessageForRemoteClient, clientIdHash, TRUE,
+                                                         &pPendingMessageQueue));
+
+                CHK(FALSE, retStatus);
+            }
+            CHK_STATUS(createSampleStreamingSession(pSampleConfiguration, pReceivedSignalingMessage->signalingMessage.peerClientId, TRUE,
+                                                    &pSampleStreamingSession));
+            freeStreamingSession = TRUE;
+            CHK_STATUS(handleOffer(pSampleConfiguration, pSampleStreamingSession, &pReceivedSignalingMessage->signalingMessage));
+            CHK_STATUS(hashTablePut(pSampleConfiguration->pRtcPeerConnectionForRemoteClient, clientIdHash, (UINT64) pSampleStreamingSession));
+
+            // If there are any ice candidate messages in the queue for this client id, submit them now.
+            CHK_STATUS(getPendingMessageQueueForHash(pSampleConfiguration->pPendingSignalingMessageForRemoteClient, clientIdHash, TRUE,
+                                                     &pPendingMessageQueue));
+            if (pPendingMessageQueue != NULL) {
+                CHK_STATUS(submitPendingIceCandidate(pPendingMessageQueue, pSampleStreamingSession));
+
+                // NULL the pointer to avoid it being freed in the cleanup
+                pPendingMessageQueue = NULL;
+            }
+
+            MUTEX_LOCK(pSampleConfiguration->streamingSessionListReadLock);
+            pSampleConfiguration->sampleStreamingSessionList[pSampleConfiguration->streamingSessionCount++] = pSampleStreamingSession;
+            MUTEX_UNLOCK(pSampleConfiguration->streamingSessionListReadLock);
+            freeStreamingSession = FALSE;
+
+            startStats = pSampleConfiguration->iceCandidatePairStatsTimerId == MAX_UINT32;
+            break;
+
+        case SIGNALING_MESSAGE_TYPE_ANSWER:
+            /*
+             * for viewer, pSampleStreamingSession should've already been created. insert the client id and
+             * streaming session into pRtcPeerConnectionForRemoteClient for subsequent ice candidate messages.
+             * Lastly check if there is any ice candidate messages queued in pPendingSignalingMessageForRemoteClient.
+             * If so then submit all of them.
+             */
+            pSampleStreamingSession = pSampleConfiguration->sampleStreamingSessionList[0];
+            CHK_STATUS(handleAnswer(pSampleConfiguration, pSampleStreamingSession, &pReceivedSignalingMessage->signalingMessage));
+            CHK_STATUS(hashTablePut(pSampleConfiguration->pRtcPeerConnectionForRemoteClient, clientIdHash, (UINT64) pSampleStreamingSession));
+
+            // If there are any ice candidate messages in the queue for this client id, submit them now.
+            CHK_STATUS(getPendingMessageQueueForHash(pSampleConfiguration->pPendingSignalingMessageForRemoteClient, clientIdHash, TRUE,
+                                                     &pPendingMessageQueue));
+            if (pPendingMessageQueue != NULL) {
+                CHK_STATUS(submitPendingIceCandidate(pPendingMessageQueue, pSampleStreamingSession));
+
+                // NULL the pointer to avoid it being freed in the cleanup
+                pPendingMessageQueue = NULL;
+            }
+
+            startStats = pSampleConfiguration->iceCandidatePairStatsTimerId == MAX_UINT32;
+            CHK_STATUS(signalingClientGetMetrics(pSampleConfiguration->signalingClientHandle, &pSampleConfiguration->signalingClientMetrics));
+            DLOGP("[Signaling offer sent to answer received time] %" PRIu64 " ms",
+                  pSampleConfiguration->signalingClientMetrics.signalingClientStats.offerToAnswerTime);
+            break;
+
+        case SIGNALING_MESSAGE_TYPE_ICE_CANDIDATE:
+            /*
+             * if peer connection hasn't been created, create an queue to store the ice candidate message. Otherwise
+             * submit the signaling message into the corresponding streaming session.
+             */
+            if (!peerConnectionFound) {
+                CHK_STATUS(getPendingMessageQueueForHash(pSampleConfiguration->pPendingSignalingMessageForRemoteClient, clientIdHash, FALSE,
+                                                         &pPendingMessageQueue));
+                if (pPendingMessageQueue == NULL) {
+                    CHK_STATUS(createMessageQueue(clientIdHash, &pPendingMessageQueue));
+                    CHK_STATUS(stackQueueEnqueue(pSampleConfiguration->pPendingSignalingMessageForRemoteClient, (UINT64) pPendingMessageQueue));
+                }
+
+                pReceivedSignalingMessageCopy = (PReceivedSignalingMessage) MEMCALLOC(1, SIZEOF(ReceivedSignalingMessage));
+
+                *pReceivedSignalingMessageCopy = *pReceivedSignalingMessage;
+
+                CHK_STATUS(stackQueueEnqueue(pPendingMessageQueue->messageQueue, (UINT64) pReceivedSignalingMessageCopy));
+
+                // NULL the pointers to not free any longer
+                pPendingMessageQueue = NULL;
+                pReceivedSignalingMessageCopy = NULL;
+            } else {
+                CHK_STATUS(handleRemoteCandidate(pSampleStreamingSession, &pReceivedSignalingMessage->signalingMessage));
+            }
+            break;
+
+        default:
+            DLOGD("Unhandled signaling message type %u", pReceivedSignalingMessage->signalingMessage.messageType);
+            break;
+    }
+
+    MUTEX_UNLOCK(pSampleConfiguration->sampleConfigurationObjLock);
+    locked = FALSE;
+
+    if (pSampleConfiguration->enableIceStats && startStats &&
+        STATUS_FAILED(retStatus = timerQueueAddTimer(pSampleConfiguration->timerQueueHandle, SAMPLE_STATS_DURATION, SAMPLE_STATS_DURATION,
+                                                     getIceCandidatePairStatsCallback, (UINT64) pSampleConfiguration,
+                                                     &pSampleConfiguration->iceCandidatePairStatsTimerId))) {
+        DLOGW("Failed to add getIceCandidatePairStatsCallback to add to timer queue (code 0x%08x). "
+              "Cannot pull ice candidate pair metrics periodically",
+              retStatus);
+
+        // Reset the returned status
+        retStatus = STATUS_SUCCESS;
+    }
+
+CleanUp:
+
+    SAFE_MEMFREE(pReceivedSignalingMessageCopy);
+    if (pPendingMessageQueue != NULL) {
+        freeMessageQueue(pPendingMessageQueue);
+    }
+
+    if (freeStreamingSession && pSampleStreamingSession != NULL) {
+        freeSampleStreamingSession(&pSampleStreamingSession);
+    }
+
+    if (locked) {
+        MUTEX_UNLOCK(pSampleConfiguration->sampleConfigurationObjLock);
+    }
+
+    CHK_LOG_ERR(retStatus);
+    return retStatus;
+}
+
+STATUS handleRemoteCandidate(PSampleStreamingSession pSampleStreamingSession, PSignalingMessage pSignalingMessage)
+{
+    STATUS retStatus = STATUS_SUCCESS;
+    RtcIceCandidateInit iceCandidate;
+    CHK(pSampleStreamingSession != NULL && pSignalingMessage != NULL, STATUS_NULL_ARG);
+
+    CHK_STATUS(deserializeRtcIceCandidateInit(pSignalingMessage->payload, pSignalingMessage->payloadLen, &iceCandidate));
+    CHK_STATUS(addIceCandidate(pSampleStreamingSession->pPeerConnection, iceCandidate.candidate));
+
+CleanUp:
+
+    CHK_LOG_ERR(retStatus);
+    return retStatus;
+}
